fix: ofga large number of streams (#2695)
