+++ src/common/infra/mod.rs
@@ -16,6 +16,7 @@
 pub mod cluster;
 pub mod config;
 pub mod wal;
+pub mod ofga;
 
 pub async fn init() -> Result<(), anyhow::Error> {
     wal::init().await?;
