+++ src/job/mod.rs
@@ -17,18 +17,8 @@ use config::{cluster, ider, utils::file::clean_empty_dirs, CONFIG, INSTANCE_ID};
 use infra::file_list as infra_file_list;
 #[cfg(feature = "enterprise")]
 use o2_enterprise::enterprise::common::infra::config::O2_CONFIG;
-#[cfg(feature = "enterprise")]
-use o2_enterprise::enterprise::openfga::authorizer::authz::{
-    get_org_creation_tuples, get_user_role_tuple, update_tuples,
-};
-#[cfg(feature = "enterprise")]
-use o2_enterprise::enterprise::openfga::meta::mapping::{NON_OWNING_ORG, OFGA_MODELS};
 use regex::Regex;
 
-#[cfg(feature = "enterprise")]
-use crate::common::infra::config::{STREAM_SCHEMAS, USERS};
-#[cfg(feature = "enterprise")]
-use crate::common::meta::user::UserRole;
 use crate::{
     common::{
         infra::config::SYSLOG_ENABLED,
@@ -124,6 +114,8 @@ pub async fn init() -> Result<(), anyhow::Error> {
     tokio::task::spawn(async move { db::alerts::watch().await });
     tokio::task::spawn(async move { db::alerts::triggers::watch().await });
     tokio::task::spawn(async move { db::organization::watch().await });
+    #[cfg(feature = "enterprise")]
+    tokio::task::spawn(async move { db::ofga::watch().await });
     tokio::task::yield_now().await; // yield let other tasks run
 
     // cache core metadata
@@ -182,6 +174,9 @@ pub async fn init() -> Result<(), anyhow::Error> {
         .await
         .expect("Load stream stats failed");
 
+    #[cfg(feature = "enterprise")]
+    db::ofga::cache().await.expect("ofga model cache failed");
+
     // check wal directory
     if cluster::is_ingester(&cluster::LOCAL_NODE_ROLE) {
         // create wal dir
@@ -206,90 +201,7 @@ pub async fn init() -> Result<(), anyhow::Error> {
     // RBAC model
     #[cfg(feature = "enterprise")]
     if O2_CONFIG.openfga.enabled {
-        let mut migrate_native_objects = false;
-        let existing_meta = match db::get_ofga_model().await {
-            Ok(Some(model)) => Some(model),
-            Ok(None) | Err(_) => {
-                migrate_native_objects = true;
-                None
-            }
-        };
-        match db::set_ofga_model(existing_meta).await {
-            Ok(store_id) => {
-                if store_id.is_empty() {
-                    log::error!("OFGA store id is empty");
-                }
-                o2_enterprise::enterprise::common::infra::config::OFGA_STORE_ID
-                    .insert("store_id".to_owned(), store_id);
-
-                if migrate_native_objects {
-                    let mut tuples = vec![];
-                    let r = STREAM_SCHEMAS.read().await;
-                    for key in r.keys() {
-                        if !key.contains('/') {
-                            continue;
-                        }
-                        let org_name = key.split('/').collect::<Vec<&str>>()[0];
-                        get_org_creation_tuples(
-                            org_name,
-                            &mut tuples,
-                            OFGA_MODELS
-                                .iter()
-                                .map(|(_, fga_entity)| fga_entity.key)
-                                .collect(),
-                            NON_OWNING_ORG.to_vec(),
-                        )
-                        .await;
-                    }
-                    drop(r);
-                    // No Data Ingested hence STREAM_SCHEMAS is empty , so we need to create at
-                    // least default org
-                    if tuples.is_empty() {
-                        get_org_creation_tuples(
-                            DEFAULT_ORG,
-                            &mut tuples,
-                            OFGA_MODELS
-                                .iter()
-                                .map(|(_, fga_entity)| fga_entity.key)
-                                .collect(),
-                            NON_OWNING_ORG.to_vec(),
-                        )
-                        .await;
-                    }
-
-                    for user_key_val in USERS.iter() {
-                        let user = user_key_val.value();
-                        if user.is_external {
-                            continue;
-                        } else {
-                            let role = if user.role.eq(&UserRole::Root) {
-                                UserRole::Admin.to_string()
-                            } else {
-                                user.role.to_string()
-                            };
-
-                            get_user_role_tuple(&role, &user.email, &user.org, &mut tuples);
-                        }
-                    }
-
-                    if tuples.is_empty() {
-                        log::info!("No orgs to update to the openfga");
-                    } else {
-                        match update_tuples(tuples, vec![]).await {
-                            Ok(_) => {
-                                log::info!("Orgs updated to the openfga");
-                            }
-                            Err(e) => {
-                                log::error!("Error updating orgs to the openfga: {}", e);
-                            }
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                log::error!("Error setting OFGA model: {:?}", e);
-            }
-        }
+        crate::common::infra::ofga::init().await;
     }
 
     // Shouldn't serve request until initialization finishes
